msgpack_django/requirements.txt
asgiref==3.11.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-msgpack==1.0.2
exceptiongroup==1.3.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.5
granian==2.6.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.42.5
locust-cloud==1.29.2
MarkupSafe==3.0.3
msgpack==1.1.2
msgpack-python==0.5.6
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
psutil==7.1.3
Pygments==2.19.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.15.0
pyzmq==27.1.0
requests==2.32.4
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
uvloop==0.22.1
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
